requirements/requirements.txt
fastapi
pydantic
pandas
numpy==1.23.5
dill
mlflow
scikit-learn==1.3.0
uvicorn
hydra-core
joblib
python-dotenv
omegaconf
python-multipart
pytest
httpx
locust
coverage
bandit
ruff
pyyaml